319 商城业务-秒杀服务-查询秒杀商品

+++ gulimall-seckill/src/main/java/com/atguigu/gulimall/seckill/vo/SeckillSkuVo.java
@@ -35,11 +35,11 @@ public class SeckillSkuVo {
     /**
      * 秒杀总量
      */
-    private Integer seckillCount;
+    private BigDecimal seckillCount;
     /**
      * 每人限购数量
      */
-    private Integer seckillLimit;
+    private BigDecimal seckillLimit;
     /**
      * 排序
      */
